requirements.txt
flask
flask_restful
flask-restful-swagger
werkzeug
pillow
